Amending Simpson's logic.
This is a recitation, and explanation, of Danny's method, which is *too good to
not steal*. Seriously, I'm impressed. I also think, generally, my recursions are
looking a little "iterative". For example, in the (simpsons) procedure I've
deleted, I iterated over the [0, n] intervals, and not through the [a, b] range
by appending 2h slices. It's... a small thing; is it? But I think the latter
style of recursion is 1) more sophisticated, 2) more idiomatic in Scheme/Racket,
and 3) indicative of greater skill.